+++ src/library/adash-messages.ads
@@ -262,6 +262,7 @@ package Adash.Messages is
       Msg_Statement_Among_Declarations,
       Msg_String_Index_Malformed,
       Msg_Not_Taken_Apart,
+      Msg_No_Such_Slice,
       Msg_Case_Not_Discrete,
       Msg_Case_Choice_Not_Static,
       Msg_Case_Choice_Covered_Twice,